Restructured into modules for easier extension (configitems, faq, etc)
Made numerous updates to restructure this to allow adding further web
service modules (GenericConfigItemConnectorSOAP, the FAQ connector, etc)
